backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlmodel==0.0.8
sqlalchemy==1.4.41
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
langchain==0.0.340
langchain-core==0.1.15
langchain-google-genai==0.0.5
google-generativeai==0.3.2
python-dotenv==1.0.0
tavily-python==0.2.6
youtube-search-python==1.6.6
youtube-transcript-api==0.5.0
requests==2.31.0
PyJWT==2.8.0
pytest==7.4.2
psutil==5.9.6
pytest-asyncio==0.21.0
alembic==1.11.1
redis==5.2.0
reportlab==4.0.4
pinecone-client==2.2.4